Game Board and Styling
For each element in boardArray, a new div is created using document.createElement(div)
The classList.add(board-cell) method adds a class to each cell div for CSS styling
setAttribute(data-index, index) adds a data-index attribute to each div element. To track the position of each cell
textContent = element sets the content of each cell.
appendChild adds the div to the boardContainerEl.

Styled the board container with grid
Flex for the board cells with width/height and colour