Computer moves/win check
getBoardState(): Retrieves the current state of the board in an array, where each cell is marked with either X, O, or an empty string.
checkWin(): A helper function that checks if a player is one move away from winning. It takes the current board state and a player marker (either X or O). It checks the board using the win combinations and finds any lines where the player is about to win.

computerMove()
first check if the current move can win
if true then checks if the human player can win then blocks the move
if no win or blocking move can be found then a random move is made